Cleaned up terraform and fixed backends
Refactored all templators to use a config holder so we can extend them in the future if we need more data.
Modified names of things in TF to reduce conflicts within the same AWS account.
Fixed S3 backend. Wasn't working at all because it can't live in a module, and can't contain string interpolation.

Still more to do to make everything fully repeatable, will probably do it in another ticket.