Fix make build-docker-local
This commit addresses a couple problems with the Dockerfile which was
resulting in a failed build. The likely culprit of this not being caught
earlier was some cached images. This does two things

1. It removes unnecessary references to things that aren't being used
any more. The ENV setting was removed in a previous commit, but not the
reference of it later in the Dockerfile. If this is in error please let
me know!
2. The go modules are cached in a layer as these are much less likely to
change between build runs than the source code itself. The technique is
simple enough; just copy the files in a separate COPY and run `go mod
download`. This reduces the hydrated build time (previous layers cached)
from  2m35s on average (3 runs) to 34s on average (3 runs).